SP:: initial testing/done

+++ contracts/chainlink/ChainlinkPriceFeed.sol
@@ -9,21 +9,19 @@ contract ChainlinkPriceFeed {
     address public DAI_FEED = 0xAed0c38402a5d19df6E4c03F4E2DceD6e29c1ee9;
 
     function getUSDCPrice() external view returns (uint256 price) {
-        (, int p,,,) = AggregatorV3Interface(USDC_FEED).latestRoundData();
+        (, int256 p,,,) = AggregatorV3Interface(USDC_FEED).latestRoundData();
         price = uint256(p);
     }
 
     function getUSDTPrice() external view returns (uint256 price) {
-        (, int p,,,) = AggregatorV3Interface(USDT_FEED).latestRoundData();
+        (, int256 p,,,) = AggregatorV3Interface(USDT_FEED).latestRoundData();
         price = uint256(p);
     }
 
     function getAverageAssetsPrice() external view returns (uint256 price) {
-
-        (, int price1,,,) = AggregatorV3Interface(USDT_FEED).latestRoundData();
-        (, int price2,,,) = AggregatorV3Interface(USDC_FEED).latestRoundData();
-        (, int price3,,,) = AggregatorV3Interface(DAI_FEED).latestRoundData();
+        (, int256 price1,,,) = AggregatorV3Interface(USDT_FEED).latestRoundData();
+        (, int256 price2,,,) = AggregatorV3Interface(USDC_FEED).latestRoundData();
+        (, int256 price3,,,) = AggregatorV3Interface(DAI_FEED).latestRoundData();
         price = uint256((price1 + price2 + price3) / 3);
-
     }
 }
